Change the version of the solidity to 0.8.7
The reason for this change is that when moving over to the local environment, the version of solidity used with the library solc is fixed for 0.8.7. So if someone were to copy and paste the code, it wouldn't work. By changing it to 0.8.7, it will work with the fixed compiler.